my_requirements.txt
absl-py==0.12.0
actionlib==1.14.0
addict==2.4.0
aiofiles==22.1.0
aiosqlite==0.20.0
angles==1.9.13
anyio==4.3.0
argon2-cffi==20.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
async-generator==1.10
attrs==20.3.0
Babel==2.14.0
backcall==0.2.0
base_local_planner==1.17.3
beautifulsoup4==4.12.3
bleach==3.3.0
blinker==1.7.0
bondpy==1.8.6
cachetools==4.2.1
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==0.4.23
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
ConfigArgParse==1.7
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
cycler==0.10.0
Cython==0.29.22
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses==0.6
debugpy==1.8.1
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distlib==0.3.8
distro==1.9.0
docutils==0.17
dynamic-reconfigure==1.7.3
entrypoints==0.3
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.3
Flask==3.0.3
freetype-py==2.2.0
fsspec==2024.2.0
future==0.18.2
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
google-auth==1.28.0
google-auth-oauthlib==0.4.4
grpcio==1.36.1
idna==2.10
image-geometry==1.16.2
imageio==2.9.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
install==1.3.5
interactive-markers==1.12.0
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==8.1.2
itsdangerous==2.2.0
jedi==0.18.0
Jinja2==3.1.3
joblib==1.0.1
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
json5==0.9.24
jsonschema==3.2.0
jsonschema-specifications==2023.12.1
jupyter-client==6.1.12
jupyter-events==0.10.0
jupyter-ydoc==0.2.5
jupyter_core==5.7.2
jupyter_packaging==0.12.3
jupyter_server==2.13.0
jupyter_server_fileid==0.9.1
jupyter_server_terminals==0.5.3
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.7
jupyterlab-pygments==0.1.2
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
kiwisolver==1.3.1
laser_geometry==1.6.7
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.1.5
matplotlib==3.4.1
matplotlib-inline==0.1.6
message-filters==1.16.0
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.2.0
mkl-service==2.3.0
mpi4py==3.1.6
mpmath==1.3.0
nbclassic==1.0.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.10.4
nest-asyncio==1.5.1
networkx==2.2
notebook==6.3.0
notebook_shim==0.2.4
numpy==1.22.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
oauthlib==3.1.0
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
open3d==0.12.0
opencv-contrib-python==4.6.0.66
overrides==7.7.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
plotly==5.21.0
plyfile==0.7.3
prometheus-client==0.10.0
prompt-toolkit==3.0.18
protobuf==3.15.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==2.7.9
pycollada==0.6
pycparser==2.20
pyglet==1.5.15
Pygments==2.8.1
pykdtree==1.3.7.post0
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.7
pyparsing==2.4.7
pyquaternion==0.9.9
pyrealsense2==2.55.1.6486
pyrender==0.1.45
pyrsistent==0.17.3
pyserial==3.5
python-dateutil==2.8.1
python-json-logger==2.0.7
python-qt-binding==0.4.4
pytorch-ignite==0.5.0.post2
pytransform3d==3.5.0
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
referencing==0.34.0
requests==2.25.1
requests-oauthlib==1.3.0
resource_retriever==1.12.7
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospkg==1.5.1
rospy==1.16.0
rosserial_python==0.9.2
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rpds-py==0.18.0
rqt-console==0.4.12
rqt-image-view==0.4.17
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-tf-tree==0.6.4
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rsa==4.7.2
rtabmap-python==0.21.4
rviz==1.14.20
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.2
Send2Trash==1.5.0
sensor-msgs==1.13.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
sklearn==0.0
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
terminado==0.9.4
testpath==0.4.4
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
threadpoolctl==2.1.0
tifffile==2021.3.31
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.4
topic-tools==1.16.0
torch==2.3.0+cu118
torch_scatter==2.1.2+pt23cu118
torchaudio==2.3.0+cu118
torchvision==0.8.2
tornado==6.1
tqdm==4.59.0
traitlets==5.14.3
trimesh==3.9.10
triton==2.3.0
turtlebot3_teleop==1.2.5
typing_extensions==4.10.0
tzdata==2024.1
urdfpy==0.0.22
urllib3==1.26.4
# Editable install with no version control (vgn==0.1.0)
-e /home/eric/Grasp_detection_GIGA/src
virtualenv==20.25.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.2
widgetsnbextension==4.0.10
xacro==1.14.17
y-py==0.6.2
ypy-websocket==0.8.4
zipp==3.18.1
